+++ crates/web-sys/src/features/gen_VideoFrameInit.rs
@@ -167,7 +167,7 @@ impl VideoFrameInit {
         self
     }
     #[deprecated = "Use `set_duration()` instead."]
-    pub fn duration(&mut self, val: f64) -> &mut Self {
+    pub fn duration(&mut self, val: u32) -> &mut Self {
         self.set_duration(val);
         self
     }
@@ -191,7 +191,7 @@ impl VideoFrameInit {
         self
     }
     #[deprecated = "Use `set_timestamp()` instead."]
-    pub fn timestamp(&mut self, val: f64) -> &mut Self {
+    pub fn timestamp(&mut self, val: i32) -> &mut Self {
         self.set_timestamp(val);
         self
     }
